# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mango",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/investimentos/serie": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Serie",
        "operationId": "serie_api_investimentos_serie_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fim"
            }
          },
          {
            "name": "recorte",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "todos",
                "renda_fixa",
                "renda_variavel"
              ],
              "type": "string",
              "default": "todos",
              "title": "Recorte"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subtype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarteiraSerie"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/posicao/transacoes": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Transacoes Posicao",
        "operationId": "transacoes_posicao_api_investimentos_posicao_transacoes_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 200,
              "title": "Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestimentoTransacaoRead"
                  },
                  "title": "Response Transacoes Posicao Api Investimentos Posicao Transacoes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/posicao/serie": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Serie Posicao",
        "operationId": "serie_posicao_api_investimentos_posicao_serie_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fim"
            }
          },
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 200,
              "title": "Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarteiraSerie"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/posicao/proventos": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Proventos Posicao",
        "operationId": "proventos_posicao_api_investimentos_posicao_proventos_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fim"
            }
          },
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 200,
              "title": "Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProventosFII"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/posicao/fundamentos": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Fundamentos Posicao",
        "operationId": "fundamentos_posicao_api_investimentos_posicao_fundamentos_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 200,
              "title": "Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundamentosFII"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/posicao/cota-serie": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Cota Serie Posicao",
        "operationId": "cota_serie_posicao_api_investimentos_posicao_cota_serie_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fim"
            }
          },
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "minItems": 1,
              "maxItems": 200,
              "title": "Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CotaSeriePonto"
                  },
                  "title": "Response Cota Serie Posicao Api Investimentos Posicao Cota Serie Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/{investimento_id}/aportes": {
      "post": {
        "tags": [
          "investimento"
        ],
        "summary": "Criar Aporte",
        "description": "Adiciona um aporte (compra) à mão a uma posição — entra no custo médio (§4.9).",
        "operationId": "criar_aporte_api_investimentos__investimento_id__aportes_post",
        "parameters": [
          {
            "name": "investimento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investimento Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AporteManualCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoTransacaoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/aportes/{aporte_id}": {
      "patch": {
        "tags": [
          "investimento"
        ],
        "summary": "Editar Aporte",
        "operationId": "editar_aporte_api_investimentos_aportes__aporte_id__patch",
        "parameters": [
          {
            "name": "aporte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Aporte Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AporteManualUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoTransacaoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "investimento"
        ],
        "summary": "Excluir Aporte",
        "operationId": "excluir_aporte_api_investimentos_aportes__aporte_id__delete",
        "parameters": [
          {
            "name": "aporte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Aporte Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/{investimento_id}": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Obter",
        "operationId": "obter_api_investimentos__investimento_id__get",
        "parameters": [
          {
            "name": "investimento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investimento Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "investimento"
        ],
        "summary": "Atualizar",
        "description": "Campos do usuário na posição: objetivo (#4), ativo/agrupamento (§4.9) e custo manual\n(valor investido quando o Pluggy não fornece). Patch parcial (`exclude_unset`) — mexer só\nnum campo não zera os outros. Posse validada (S3).",
        "operationId": "atualizar_api_investimentos__investimento_id__patch",
        "parameters": [
          {
            "name": "investimento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investimento Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestimentoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/{investimento_id}/transacoes": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Listar Transacoes",
        "operationId": "listar_transacoes_api_investimentos__investimento_id__transacoes_get",
        "parameters": [
          {
            "name": "investimento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investimento Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestimentoTransacaoRead"
                  },
                  "title": "Response Listar Transacoes Api Investimentos  Investimento Id  Transacoes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investimentos/{investimento_id}/proventos": {
      "get": {
        "tags": [
          "investimento"
        ],
        "summary": "Proventos",
        "operationId": "proventos_api_investimentos__investimento_id__proventos_get",
        "parameters": [
          {
            "name": "investimento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investimento Id"
            }
          },
          {
            "name": "inicio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fim"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProventosFII"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AporteManualCreate": {
        "properties": {
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "quantidade": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantidade"
          },
          "valor_centavos": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Valor Centavos"
          }
        },
        "type": "object",
        "required": [
          "data",
          "quantidade",
          "valor_centavos"
        ],
        "title": "AporteManualCreate",
        "description": "Aporte (compra) informado à mão, quando o banco não traz o histórico completo (§4.9)."
      },
...
      "CarteiraSerie": {
        "properties": {
          "recorte": {
            "type": "string",
            "title": "Recorte"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "pontos": {
            "items": {
              "$ref": "#/components/schemas/CarteiraSeriePonto"
            },
            "type": "array",
            "title": "Pontos"
          },
          "reconstruido_ate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconstruido Ate"
          }
        },
        "type": "object",
        "required": [
          "recorte",
          "subtype",
          "pontos"
        ],
        "title": "CarteiraSerie"
      },
...
      "CotaSeriePonto": {
        "properties": {
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "valor_centavos": {
            "type": "integer",
            "title": "Valor Centavos"
          }
        },
        "type": "object",
        "required": [
          "data",
          "valor_centavos"
        ],
        "title": "CotaSeriePonto"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestimentoTransacaoRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "investimento_id": {
            "type": "integer",
            "title": "Investimento Id"
          },
          "pluggy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pluggy Id"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "movement_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movement Type"
          },
          "amount_centavos": {
            "type": "integer",
            "title": "Amount Centavos"
          },
          "value_unitario": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Unitario"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "net_amount_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Amount Centavos"
          },
          "expenses_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expenses Centavos"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "brokerage_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brokerage Number"
          },
          "manual": {
            "type": "boolean",
            "title": "Manual"
          }
        },
        "type": "object",
        "required": [
          "id",
          "investimento_id",
          "amount_centavos",
          "manual"
        ],
        "title": "InvestimentoTransacaoRead"
      },
...
      "ProventosFII": {
        "properties": {
          "investimento_id": {
            "type": "integer",
            "title": "Investimento Id"
          },
          "inicio": {
            "type": "string",
            "format": "date",
            "title": "Inicio"
          },
          "fim": {
            "type": "string",
            "format": "date",
            "title": "Fim"
          },
          "total_centavos": {
            "type": "integer",
            "title": "Total Centavos"
          },
          "total_isento_centavos": {
            "type": "integer",
            "title": "Total Isento Centavos"
          },
          "dy_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dy Pct"
          },
          "proventos": {
            "items": {
              "$ref": "#/components/schemas/InvestimentoTransacaoRead"
            },
            "type": "array",
            "title": "Proventos"
          }
        },
        "type": "object",
        "required": [
          "investimento_id",
          "inicio",
          "fim",
          "total_centavos",
          "total_isento_centavos",
          "dy_pct",
          "proventos"
        ],
        "title": "ProventosFII"
      },
...
    }
  }
}